# optimize, simplify IP.Next, IP.Prior