… inlining_test.go