```
name           old time/op  new time/op  delta
IPNextPrior-4  4.18µs ± 3%  0.34µs ±12%  -91.86%  (p=0.008 n=5+5)
```